self destructive reduction

+++ phalanx/src/main/java/org/smallmind/phalanx/wire/InOut.java
@@ -42,5 +42,5 @@ import java.lang.annotation.Target;
 public @interface InOut {
 
   // Any positive value (> 0) will override the default service timeout seconds
-  int timeoutSeconds ();
+  long timeoutSeconds ();
 }
