+++ phalanx/src/main/java/org/smallmind/phalanx/wire/Whisper.java
@@ -42,5 +42,5 @@ import java.lang.annotation.Target;
 public @interface Whisper {
 
   // Any positive value (> 0) will override the default service timeout seconds
-  int timeoutSeconds () default 0;
+  long timeoutSeconds () default 0;
 }
